Refactored vue-cli-plugin tests.
