+++ packages/docs/pages/docs/introduction/installation/nuxt.md
