+++ packages/vue-cli-plugin/tests/app-ts/srcset/App.vue
