+++ packages/vue-cli-plugin/tests/app-ts/srcset/customizable/src/main.ts
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+import App from '../../App.vue';
+
+Vue.config.productionTip = false;
+
+new Vue({
+    render: h => h(App),
+}).$mount('#app');
