+++ packages/vue-cli-plugin/tests/app-ts/srcset/default/src/main.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import App from './App.vue';
+import App from '../App.vue';
 
 Vue.config.productionTip = false;
