+++ packages/vue-cli-plugin/tests/app-ts/srcset/shims-tsx.d.ts
