+++ packages/vue-cli-plugin/tests/app-ts/srcset/shims-vue.d.ts
