+++ packages/vue-cli-plugin/tests/app/srcset/App.vue
